server/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
google-generativeai>=0.8
newsapi-python==0.2.7
azure-ai-inference
requests
beautifulsoup4
pandas
lxml
scikit-learn
gunicorn
pinecone
pypdf